AITuber/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "aituber-orchestrator"
version = "0.1.0"
description = "AITuber Orchestrator – YouTube streaming host brain"
requires-python = ">=3.11"
dependencies = [
    "websockets>=12.0,<14",
    "google-api-python-client>=2.100,<3",
    "google-auth-oauthlib>=1.1,<2",
    "openai>=1.10,<2",
    "numpy>=1.26,<3",
    "aiohttp>=3.9,<4",
    "psutil>=5.9,<7",
    "python-dotenv>=1.0,<2",
    "sounddevice>=0.4,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "ruff>=0.4,<1",
    "black>=24,<26",
]

[project.scripts]
aituber = "orchestrator.main:main"

[tool.setuptools.packages.find]
include = ["orchestrator*"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.black]
target-version = ["py311"]
line-length = 99

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
